refactor(workspace): extract quspin-types crate
Split foundation types out of quspin-core into a new quspin-types crate.
First of seven PRs splitting the monolithic quspin-core into a layered
workspace per docs/superpowers/specs/2026-04-18-crate-split-design.md.

Moves (verbatim, no logic changes):
- error.rs, primitive.rs, dtype.rs, expm/compute.rs → quspin-types/src/
- linear_operator/{mod.rs, fn_op.rs} → quspin-types/src/linear_operator/

quspin-core now depends on quspin-types and re-exports its modules
(error, primitive, dtype, compute, linear_operator_types) at the crate
root. linear_operator/mod.rs and expm/compute.rs are kept as thin
forwarding shims so existing crate-local paths (e.g. super::compute,
crate::linear_operator::QMatrixOperator) continue to resolve while
QMatrix lives in quspin-core.

quspin-py is untouched: quspin_core::error::QuSpinError and friends
still resolve through the re-exports.
