+++ crates/quspin-core/Cargo.toml
@@ -7,6 +7,7 @@ edition.workspace = true
 large-int = []
 
 [dependencies]
+quspin-types = { path = "../quspin-types" }
 num-complex = { workspace = true }
 ndarray = { workspace = true }
 rayon = { workspace = true }
