+++ crates/quspin-core/src/lib.rs
@@ -1,21 +1,24 @@
+pub use quspin_types::{
+    compute, dtype, error, linear_operator as linear_operator_types, primitive,
+};
+
 pub mod basis;
 pub mod bitbasis;
-pub mod dtype;
-pub mod error;
 pub mod expm;
 pub mod hamiltonian;
 pub mod krylov;
 pub mod linear_operator;
 pub mod operator;
-pub mod primitive;
 pub mod qmatrix;
 
-pub use error::QuSpinError;
 pub use expm::{
-    ExpmComputation, expm_multiply, expm_multiply_auto, expm_multiply_auto_into,
-    expm_multiply_many, expm_multiply_many_auto, expm_multiply_many_auto_into,
+    expm_multiply, expm_multiply_auto, expm_multiply_auto_into, expm_multiply_many,
+    expm_multiply_many_auto, expm_multiply_many_auto_into,
 };
 pub use hamiltonian::{Hamiltonian, HamiltonianInner, IntoHamiltonianInner, SchrodingerEq};
-pub use linear_operator::{DynLinearOperator, FnLinearOperator, LinearOperator, QMatrixOperator};
+pub use linear_operator::{LinearOperator, QMatrixOperator};
 pub use operator::{Operator, ParseOp};
-pub use primitive::Primitive;
+pub use quspin_types::{
+    AtomicAccum, CIndexDType, DynLinearOperator, ExpmComputation, FnLinearOperator,
+    FnLinearOperatorBuilder, Primitive, QuSpinError, ValueDType,
+};
