+++ crates/quspin-core/src/operator/bond/operator.rs
@@ -1,9 +1,9 @@
 use crate::bitbasis::BitInt;
 use crate::bitbasis::manip::{DitManip, DynamicDitManip};
-use crate::error::QuSpinError;
 use crate::operator::Operator;
 use ndarray::Array2;
 use num_complex::Complex;
+use quspin_types::QuSpinError;
 
 // ---------------------------------------------------------------------------
 // BondTerm
