+++ crates/quspin-core/src/operator/fermion/operator.rs
@@ -2,9 +2,9 @@ use num_complex::Complex;
 use smallvec::SmallVec;
 
 use crate::bitbasis::BitInt;
-use crate::error::QuSpinError;
 use crate::operator::Operator;
 use crate::operator::ParseOp;
+use quspin_types::QuSpinError;
 
 // ---------------------------------------------------------------------------
 // FermionOp
