+++ crates/quspin-types/src/error.rs
