+++ crates/quspin-types/src/primitive.rs
